+++ Bolt-Core/src/Core/Utils/Filesystem/Directorypath.cpp
@@ -36,7 +36,7 @@ namespace Bolt
 
 	Directorypath Directorypath::Parent() const
 	{
-		size_t index = m_Path.find_last_of('\\', m_Path.length() - 2);
+		uint32_t index = m_Path.find_last_of('\\', m_Path.length() - 2);
 		return m_Path.substr(0, index);
 	}
