+++ Bolt-Core/src/Graphics/Buffers/IndexBuffer.cpp
@@ -5,24 +5,24 @@
 namespace Bolt
 {
 
-	IndexBuffer::IndexBuffer(size_t indexCount, BufferUsage usage) : IndexBuffer(nullptr, indexCount, usage)
+	IndexBuffer::IndexBuffer(uint32_t indexCount, BufferUsage usage) : IndexBuffer(nullptr, indexCount, usage)
 	{
 
 	}
 
-	IndexBuffer::IndexBuffer(const uint* indices, size_t indexCount, BufferUsage usage) : Buffer((const void*)indices, indexCount * sizeof(uint), BufferTarget::ElementBuffer, usage)
+	IndexBuffer::IndexBuffer(const uint32_t* indices, uint32_t indexCount, BufferUsage usage) : Buffer((const void*)indices, indexCount * sizeof(uint32_t), BufferTarget::ElementBuffer, usage)
 	{
 
 	}
 
 	int IndexBuffer::IndexCount() const
 	{
-		return Size() / sizeof(uint);
+		return Size() / sizeof(uint32_t);
 	}
 
 	std::unique_ptr<IndexBuffer> IndexBuffer::Clone() const
 	{
-		std::unique_ptr<IndexBuffer> indexBuffer = std::make_unique<IndexBuffer>(Size() / sizeof(uint), Usage());
+		std::unique_ptr<IndexBuffer> indexBuffer = std::make_unique<IndexBuffer>(Size() / sizeof(uint32_t), Usage());
 		byte* data = new byte[Size()];
 		Download(data, Size(), 0);
 		indexBuffer->Upload(data, Size(), 0);
